+++ example/src/viewer.js
@@ -7,6 +7,11 @@ import fileDrop from 'file-drops';
 import fileOpen from 'file-open';
 
 import exampleXML from '../resources/example.bpmn';
+import minimapModule from 'diagram-js-minimap';
+
+import gridModule from 'diagram-js-grid';
+import download from 'downloadjs';
+import { svgToPng } from './utils';
 
 
 const url = new URL(window.location.href);
@@ -14,6 +19,8 @@ const url = new URL(window.location.href);
 const persistent = url.searchParams.has('p');
 const active = url.searchParams.has('e');
 const presentationMode = url.searchParams.has('pm');
+
+let fileName = 'diagram.bpmn';
 
 const initialDiagram = (() => {
   try {
@@ -78,7 +85,9 @@ const viewer = new BpmnViewer({
   container: '#canvas',
   additionalModules: [
     ExampleModule,
-    TokenSimulationModule
+    TokenSimulationModule,
+    gridModule,
+    minimapModule
   ]
 });
 
@@ -114,10 +123,34 @@ function openFile(files) {
 
   hideMessage();
 
+  fileName = files[0].name;
+
   openDiagram(files[0].contents);
 }
 
+function exportPNG() {
+  viewer.saveSVG().then(({ svg }) => {
+    svgToPng(svg).then(png => {
+      download(png, fileName.replace(/\.bpmn$/i, '.png'), 'image/png');
+    });
+  });
+}
+
+function exportSVG() {
+  viewer.saveSVG().then(({ svg }) => {
+    download(svg, fileName.replace(/\.bpmn$/i, '.svg'), 'image/svg+xml');
+  });
+}
+
 document.body.addEventListener('dragover', fileDrop('Open BPMN diagram', openFile), false);
+
+document.querySelector('#export-png').addEventListener('click', function(event) {
+  exportPNG();
+});
+
+document.querySelector('#export-svg').addEventListener('click', function(event) {
+  exportSVG();
+});
 
 document.body.addEventListener('keydown', function(event) {
   if (event.code === 'KeyO' && (event.metaKey || event.ctrlKey)) {
@@ -151,3 +184,7 @@ if (remoteDiagram) {
 } else {
   openDiagram(initialDiagram);
 }
+
+// expose for theming
+window.bpmnjs = viewer;
+