+++ webpack.config.js
@@ -12,7 +12,8 @@ module.exports = (env, argv) => {
     mode,
     entry: {
       viewer: './example/src/viewer.js',
-      modeler: './example/src/modeler.js'
+      modeler: './example/src/modeler.js',
+      basic: './example/basic/src/app.js'
     },
     output: {
       filename: 'dist/[name].js',
@@ -31,7 +32,11 @@ module.exports = (env, argv) => {
         patterns: [
           { from: './assets', to: 'dist/vendor/bpmn-js-token-simulation/assets' },
           { from: 'bpmn-js/dist/assets', context: 'node_modules', to: 'dist/vendor/bpmn-js/assets' },
-          { from: '@bpmn-io/properties-panel/dist/assets', context: 'node_modules', to: 'dist/vendor/bpmn-js-properties-panel/assets' }
+          { from: '@bpmn-io/properties-panel/dist/assets', context: 'node_modules', to: 'dist/vendor/bpmn-js-properties-panel/assets' },
+          { from: 'bpmn-js-bpmnlint/dist/assets', context: 'node_modules', to: 'dist/vendor/bpmn-js-bpmnlint/assets' },
+          { from: 'bpmn-js-color-picker/colors/color-picker.css', context: 'node_modules', to: 'dist/vendor/bpmn-js-color-picker/colors/color-picker.css' },
+          { from: 'diagram-js-minimap/assets/diagram-js-minimap.css', context: 'node_modules', to: 'dist/vendor/diagram-js-minimap/assets/diagram-js-minimap.css' },
+          { from: '.bpmnlintrc', to: '.bpmnlintrc', noErrorOnMissing: true }
         ]
       }),
       new DefinePlugin({